bone_web/requirements.txt
flask
torch
torchvision
timm
pillow